+++ test/pal.spec.ts
@@ -2,6 +2,7 @@
  * Tests for PAL.
  */
 import { NetworkClientFactory } from "@iota-pico/core/dist/factories/networkClientFactory";
+import { PlatformCryptoFactory } from "@iota-pico/core/dist/factories/platformCryptoFactory";
 import { RngServiceFactory } from "@iota-pico/core/dist/factories/rngServiceFactory";
 import { NetworkEndPoint } from "@iota-pico/core/dist/network/networkEndPoint";
 import * as chai from "chai";
@@ -22,6 +23,11 @@ describe("PAL", () => {
         it("can be called and register rng service", async () => {
             await PAL.initialize();
             chai.expect(RngServiceFactory.instance().create("default")).to.not.be.equal(undefined);
+        });
+
+        it("can be called and register platform crypto", async () => {
+            await PAL.initialize();
+            chai.expect(PlatformCryptoFactory.instance().create("default")).to.not.be.equal(undefined);
         });
 
         it("can be called twice", async () => {